Fixed marketplace-page test suite.
