+++ sonarQuest-frontend/src/app/app.module.ts
@@ -100,7 +100,7 @@ import {LoadingService} from './services/loading.service';
 import {UserToWorldService} from './services/user-to-world.service';
 import {PermissionService} from './services/permission.service';
 import {WizardService} from './services/wizard.service';
-import {ArtefactViewDetailsComponent} from './pages/marketplace-page/components/marketplace-artefact-view/marketplace-artefact-view.component';
+import {ArtefactViewDetailsComponent} from './pages/marketplace-page/components/artefact-view-details/artefact-view-details.component';
 import {EventService} from './services/event.service'
 import {EventPageComponent} from './pages/event-page/event-page.component';
 import {WebsocketService} from './services/websocket.service';
