+++ sonarQuest-frontend/src/app/pages/marketplace-page/components/artefact-view-details/artefact-view-details.component.html
