+++ sonarQuest-frontend/src/app/pages/marketplace-page/components/artefact-view-details/artefact-view-details.component.spec.ts
@@ -1,7 +1,7 @@
 import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 
-import {AvatarEditComponent} from "./my-avatar-edit.component";
-import {BrowserModule, DomSanitizer} from "@angular/platform-browser";
+import {ArtefactViewDetailsComponent} from "./artefact-view-details.component";
+import {BrowserModule} from "@angular/platform-browser";
 import {BrowserAnimationsModule} from "@angular/platform-browser/animations";
 import {RouterTestingModule} from "@angular/router/testing";
 import {FormsModule, ReactiveFormsModule} from "@angular/forms";
@@ -17,18 +17,22 @@ import {
   MatToolbarModule,
   MatTooltipModule
 } from "@angular/material";
-import {UserServiceTestingModule} from "../../../../services/user.service.mock.module";
-import {ImageService} from "../../../../services/image.service";
+import {CovalentDataTableModule} from "@covalent/core";
 
-describe('AvatarEditComponent', () => {
-  let component: AvatarEditComponent;
-  let fixture: ComponentFixture<AvatarEditComponent>;
+describe('ArtefactViewDetailsComponent', () => {
+  let component: ArtefactViewDetailsComponent;
+  let fixture: ComponentFixture<ArtefactViewDetailsComponent>;
 
-  const data = {};
+  const data = {
+    minLevel: {
+      levelNumber: 1
+    }
+
+  };
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ AvatarEditComponent ],
+      declarations: [ ArtefactViewDetailsComponent ],
       imports: [
         BrowserModule,
         BrowserAnimationsModule,
@@ -43,20 +47,18 @@ describe('AvatarEditComponent', () => {
         MatDividerModule,
         MatToolbarModule,
         MatDialogModule,
-        UserServiceTestingModule,
+        CovalentDataTableModule
       ],
       providers: [
         {provide: MAT_DIALOG_DATA, useValue: data},
-        {provide: MatDialogRef, useValue: {}},
-        ImageService,
-        DomSanitizer
+        {provide: MatDialogRef, useValue: {}}
       ]
     })
       .compileComponents();
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(AvatarEditComponent);
+    fixture = TestBed.createComponent(ArtefactViewDetailsComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
